requirements.txt
future==0.18.3
pyserial==3.5
tkintertable==1.3.3
ttkthemes~=3.2.2
pytest~=7.3.1